--- a/cpp/cpp_boggle.cc
+++ b/cpp/cpp_boggle.cc
@@ -85,7 +85,8 @@ PYBIND11_MODULE(cpp_boggle, m)
         .def("NumNodes", &Trie::NumNodes)
         .def("SetAllMarks", &Trie::SetAllMarks)
         .def_static("ReverseLookup", py::overload_cast<const Trie*, const Trie*>(&Trie::ReverseLookup))
-        .def_static("CreateFromFile", &Trie::CreateFromFile);
+        .def_static("CreateFromFile", &Trie::CreateFromFile)
+        .def_static("CreateFromFileWithGrouping", &Trie::CreateFromFileWithGrouping);
 
     declare_boggler<3, 3>(m, "Boggler33");
     declare_boggler<3, 4>(m, "Boggler34");