--- a/cpp/trie.h
+++ b/cpp/trie.h
@@ -3,6 +3,7 @@
 
 #include <memory>
 #include <string>
+#include <unordered_map>
 #include <vector>
 #include <sys/types.h>
 #include <stdint.h>
@@ -31,6 +32,7 @@ class Trie {
   // Returns a pointer to the new Trie node at the end of the word.
   Trie* AddWord(const char* wd);
   static unique_ptr<Trie> CreateFromFile(const char* filename);
+  static unique_ptr<Trie> CreateFromFileWithGrouping(const char* filename, unordered_map<char, char> letter_grouping);
 
   // Some slower methods that operate on the entire Trie (not just a node).
   size_t Size();